+++ src/pages/Login.jsx
@@ -1,0 +1,7 @@
+import LoginForm from "../component/LoginForm";
+
+function Login({ users }) {
+  return <LoginForm users={users} />;
+}
+
+export default Login;
